Crate xrpl

Utilities for interacting with the XRP Ledger.

A pure Rust implementation for interacting with the XRP Ledger. The xrpl-rust crate simplifies the hardest parts of XRP Ledger interaction including serialization and transaction signing while providing idiomatic Rust functionality for XRP Ledger transactions and core server API (rippled) objects.

Quick Start

TODO

The XRP Ledger

For the user guide and further documentation, please read XRP Ledger.

Re-exports

pub extern crate serde_json;

Modules

account

asynch

clients

constants

Collection of public constants for XRPL.

core

Core codec functions for interacting with the XRPL.

ledger

macros

models

Top-level modules for the models package.

transaction

utils

Convenience utilities for the XRP Ledger

wallet

Methods for working with XRPL wallets.

Macros

serde_with_tag

A macro to tag a struct externally. With serde attributes, unfortunately it is not possible to serialize a struct to json with its name as key and its fields as value. Example: {"Example":{"Field1":"hello","Field2":"world"}}

skip_err

Enums

XRPLSerdeJsonError