Module xrpl::core::addresscodec

This module contains commonly-used constants.

Modules

exceptions

General XRPL Address Codec Exception.

utils

This module contains commonly-used utilities.

Functions

classic_address_to_xaddress

Returns the X-Address representation of the data.

decode_account_public_key

Returns the decoded bytes of the node public key.

decode_classic_address

Returns the decoded bytes of the classic address.

decode_node_public_key

Returns the decoded bytes of the node public key.

decode_seed

Returns an encoded seed.

encode_account_public_key

Returns the account public key encoding of these bytes as a base58 string.

encode_classic_address

Returns the classic address encoding of these bytes as a base58 string.

encode_node_public_key

Returns the node public key encoding of these bytes as a base58 string.

encode_seed

Returns an encoded seed.

is_valid_classic_address

Returns whether classic_address is a valid classic address.

is_valid_xaddress

Returns whether xaddress is a valid X-Address.

xaddress_to_classic_address

Returns a tuple containing the classic address, tag, and whether the address is on a test network for an X-Address.