Module xrpl::core::binarycodec::definitions

Functions for encoding objects into the XRP Ledger’s canonical binary format and decoding them.

Re-exports

pub use self::types::*;

Modules

types

Maps and helpers providing serialization-related information about fields.

Structs

FieldHeader

A container class for simultaneous storage of a field’s type code and field code.

FieldInfo

Model object for field info metadata from the “fields” section of definitions.json.

FieldInstance

A collection of serialization information about a specific field type.

Constants

CODE_MAX_VALUE

CODE_MIN_VALUE