Module xrpl::core::binarycodec::definitions::types

Maps and helpers providing serialization-related information about fields.

Structs

DefinitionMap

Loads JSON from the definitions file and converts it to a preferred format. The definitions file contains information required for the XRP Ledger’s canonical binary serialization format.

Definitions

Field

LedgerEntryTypes

TransactionResults

=(

TransactionTypes

Types

Traits

DefinitionHandler

Functions

get_field_code

Returns the field code associated with the given field.

get_field_header_from_name

Returns a FieldHeader object for a field of the given field name.

get_field_instance

Return a FieldInstance object for the given field name.

get_field_name_from_header

Returns the field name described by the given FieldHeader object.

get_field_type_code

Returns the type code associated with the given field.

get_field_type_name

Returns the serialization data type for the given field name.

get_ledger_entry_type_code

Return an integer representing the given ledger entry type string in an enum.

get_ledger_entry_type_name

Return an integer representing the given ledger entry type string in an enum.

get_transaction_result_code

Return an integer representing the given transaction result string in an enum.

get_transaction_result_name

Return string representing the given transaction result type from the enum.

get_transaction_type_code

Return an integer representing the given transaction type string in an enum.

get_transaction_type_name

Return an integer representing the given transaction type string in an enum.

load_definition_map

Retrieve the definition map.