Module xrpl::core::keypairs

Core codec functions for interacting with the XRPL.

Re-exports

pub use self::algorithms::Ed25519;

pub use self::algorithms::Secp256k1;

Modules

algorithms

Ed25519 elliptic curve cryptography interface. SECP256K1 elliptic curve cryptography interface.

exceptions

XRPL keypair codec exceptions.

utils

Miscellaneous helper functions.

Traits

CryptoImplementation

Trait for cryptographic algorithms in the XRP Ledger. The classes for all cryptographic algorithms are derived from this trait.

Functions

derive_classic_address

Derive the XRP Ledger classic address for a given public key. For more information, see Address Derivation.

derive_keypair

Derive the public and private keys from a given seed value.

generate_seed

Generate a seed value that cryptographic keys can be derived from.

is_valid_message

Verifies the signature on a given message.

sign

Sign a message using a given private key.