Module xrpl::models::ledger::objects::directory_node

Structs

DirectoryNode

The DirectoryNode object type provides a list of links to other objects in the ledger’s state tree. A single conceptual Directory takes the form of a doubly linked list, with one or more DirectoryNode objects each containing up to 32 IDs of other objects. The first object is called the root of the directory, and all objects other than the root object can be added or deleted as necessary.