Struct xrpl::models::requests::ledger_entry::LedgerEntry

pub struct LedgerEntry<'a> {

    pub common_fields: CommonFields<'a>,
    pub account_root: Option<Cow<'a, str>>,
    pub binary: Option<bool>,
    pub check: Option<Cow<'a, str>>,
    pub deposit_preauth: Option<DepositPreauth<'a>>,
    pub directory: Option<Directory<'a>>,
    pub escrow: Option<Escrow<'a>>,
    pub index: Option<Cow<'a, str>>,
    pub ledger_lookup: Option<LookupByLedgerRequest<'a>>,
    pub offer: Option<Offer<'a>>,
    pub payment_channel: Option<Cow<'a, str>>,
    pub ripple_state: Option<RippleState<'a>>,
    pub ticket: Option<Ticket<'a>>,

}

The ledger_entry method returns a single ledger object from the XRP Ledger in its raw format. See ledger formats for information on the different types of objects you can retrieve.

See Ledger Formats: <https://xrpl.org/ledger-data-formats.html#ledger-data-formats>

See Ledger Entry: <https://xrpl.org/ledger_entry.html#ledger_entry>

Fields
common_fields: CommonFields<'a>

The common fields shared by all requests.

account_root: Option<Cow<'a, str>>binary: Option<bool>

If true, return the requested ledger object’s contents as a hex string in the XRP Ledger’s binary format. Otherwise, return data in JSON format. The default is false.

check: Option<Cow<'a, str>>deposit_preauth: Option<DepositPreauth<'a>>directory: Option<Directory<'a>>escrow: Option<Escrow<'a>>index: Option<Cow<'a, str>>ledger_lookup: Option<LookupByLedgerRequest<'a>>

The unique identifier of a ledger.

offer: Option<Offer<'a>>payment_channel: Option<Cow<'a, str>>ripple_state: Option<RippleState<'a>>ticket: Option<Ticket<'a>>
Implementations

impl<'a> LedgerEntry<'a>

pub fn new( id: Option<Cow<'a, str>>, account_root: Option<Cow<'a, str>>, binary: Option<bool>, check: Option<Cow<'a, str>>, deposit_preauth: Option<DepositPreauth<'a>>, directory: Option<Directory<'a>>, escrow: Option<Escrow<'a>>, index: Option<Cow<'a, str>>, ledger_hash: Option<Cow<'a, str>>, ledger_index: Option<LedgerIndex<'a>>, offer: Option<Offer<'a>>, payment_channel: Option<Cow<'a, str>>, ripple_state: Option<RippleState<'a>>, ticket: Option<Ticket<'a>>, ) -> Self