Struct xrpl::models::results::server_info::Info

pub struct Info<'a> {

    pub amendment_blocked: Option<bool>,
    pub build_version: Cow<'a, str>,
    pub closed_ledger: Option<LedgerInfo<'a>>,
    pub complete_ledgers: Cow<'a, str>,
    pub counters: Option<Value>,
    pub current_activity: Option<Value>,
    pub hostid: Option<Cow<'a, str>>,
    pub io_latency_ms: u32,
    pub jq_trans_overflow: Option<Cow<'a, str>>,
    pub last_close: LastClose,
    pub load: Option<Load<'a>>,
    pub load_factor: u32,
    pub load_factor_local: Option<u32>,
    pub load_factor_net: Option<u32>,
    pub load_factor_cluster: Option<u32>,
    pub load_factor_fee_escalation: Option<u32>,
    pub load_factor_fee_queue: Option<u32>,
    pub load_factor_server: Option<u32>,
    pub peers: u32,
    pub ports: Option<Cow<'a, [Value]>>,
    pub pubkey_node: Cow<'a, str>,
    pub pubkey_validator: Option<Cow<'a, str>>,
    pub reporting: Option<Reporting<'a>>,
    pub server_state: Cow<'a, str>,
    pub server_state_duration_us: Option<u64>,
    pub state_accounting: Option<Value>,
    pub time: Option<Cow<'a, str>>,
    pub uptime: Option<u64>,
    pub validated_ledger: Option<LedgerInfo<'a>>,
    pub validation_quorum: u32,
    pub validator_list_expires: Option<u32>,
    pub validator_list: Option<ValidatorList<'a>>,

}

Server information

Fields
amendment_blocked: Option<bool>

If true, this server is amendment blocked

build_version: Cow<'a, str>

The version number of the running rippled server

closed_ledger: Option<LedgerInfo<'a>>

Information about the most recently closed ledger that has not been validated

complete_ledgers: Cow<'a, str>

Range expression indicating the sequence numbers of the ledger versions in the database

counters: Option<Value>

Performance metrics for RPC calls and JobQueue

current_activity: Option<Value>

Items currently being run in the job queue

hostid: Option<Cow<'a, str>>

Server hostname or RFC-1751 word based on node public key

io_latency_ms: u32

Amount of time spent waiting for I/O operations, in milliseconds

jq_trans_overflow: Option<Cow<'a, str>>

Number of times server had over 250 transactions waiting

last_close: LastClose

Information about the last ledger close

load: Option<Load<'a>>

Detailed information about the current load state

load_factor: u32

Current transaction cost multiplier

load_factor_local: Option<u32>

Transaction cost multiplier based on local load

load_factor_net: Option<u32>

Transaction cost multiplier from network load

load_factor_cluster: Option<u32>

Transaction cost multiplier from cluster load

load_factor_fee_escalation: Option<u32>

Transaction cost multiplier for open ledger

load_factor_fee_queue: Option<u32>

Transaction cost multiplier for queue

load_factor_server: Option<u32>

Transaction cost multiplier excluding open ledger

peers: u32

Number of connected peer servers

ports: Option<Cow<'a, [Value]>>

List of ports listening for API commands

pubkey_node: Cow<'a, str>

Public key for peer-to-peer communications

pubkey_validator: Option<Cow<'a, str>>

Public key for ledger validations

reporting: Option<Reporting<'a>>

Reporting mode configuration information

server_state: Cow<'a, str>

Current server state

server_state_duration_us: Option<u64>

Microseconds in current state

state_accounting: Option<Value>

Server state accounting information

time: Option<Cow<'a, str>>

Current UTC time according to server

uptime: Option<u64>

Seconds server has been operational

validated_ledger: Option<LedgerInfo<'a>>

Information about the most recent validated ledger

validation_quorum: u32

Minimum required trusted validations

validator_list_expires: Option<u32>

Validator list expiration time

validator_list: Option<ValidatorList<'a>>

Validator list information