Trait xrpl::models::Model

pub trait Model {
    // Provided methods
    fn get_errors(&self) -> XRPLModelResult<()> { ... }
    fn validate(&self) -> XRPLModelResult<()> { ... }
    fn is_valid(&self) -> bool { ... }
}

A trait that implements basic functions to every model.

Provided Methods

fn get_errors(&self) -> XRPLModelResult<()>

Collects a models errors and returns the first error that occurs.

fn validate(&self) -> XRPLModelResult<()>

Simply forwards the error from get_errors if there was one.

fn is_valid(&self) -> bool

Checks if the model is valid.