Module xrpl::models::transactions

Modules

account_delete

account_set

amm_bid

amm_create

amm_delete

amm_deposit

amm_vote

amm_withdraw

check_cancel

check_cash

check_create

deposit_preauth

escrow_cancel

escrow_create

escrow_finish

exceptions

metadata

nftoken_accept_offer

nftoken_burn

nftoken_cancel_offer

nftoken_create_offer

nftoken_mint

offer_cancel

offer_create

payment

payment_channel_claim

payment_channel_create

payment_channel_fund

pseudo_transactions

set_regular_key

signer_list_set

ticket_create

trust_set

xchain_account_create_commit

xchain_add_account_create_attestation

xchain_add_claim_attestation

xchain_claim

xchain_commit

xchain_create_bridge

xchain_create_claim_id

xchain_modify_bridge

Structs

AuthAccount

Represents one entry in a list of AuthAccounts used in AMMBid transaction.

CommonFields

The base fields for all transaction models.

Memo

An arbitrary piece of data attached to a transaction. A transaction can have multiple Memo objects as an array in the Memos field.

PreparedTransaction

SignedTransaction

Signer

One Signer in a multi-signature. A multi-signed transaction can have an array of up to 8 Signers, each contributing a signature, in the Signers field.

Enums

Flag

TransactionType

Enum containing the different Transaction types.

Traits

Transaction

Standard functions for transactions.