Module xrpl::models::transactions::payment_channel_claim

Structs

PaymentChannelClaim

Claim XRP from a payment channel, adjust the payment channel’s expiration, or both.

PaymentChannelClaimFlagIter

An iterator over the variants of PaymentChannelClaimFlag

Enums

PaymentChannelClaimFlag

Transactions of the PaymentChannelClaim type support additional values in the Flags field. This enum represents those options.