Struct xrpl::models::transactions::CommonFields

pub struct CommonFields<'a, F>
where
    F: IntoEnumIterator + Serialize + Debug,
{

    pub account: Cow<'a, str>,
    pub transaction_type: TransactionType,
    pub account_txn_id: Option<Cow<'a, str>>,
    pub fee: Option<XRPAmount<'a>>,
    pub flags: FlagCollection<F>,
    pub last_ledger_sequence: Option<u32>,
    pub memos: Option<Vec<Memo>>,
    pub network_id: Option<u32>,
    pub sequence: Option<u32>,
    pub signers: Option<Vec<Signer>>,
    pub signing_pub_key: Option<Cow<'a, str>>,
    pub source_tag: Option<u32>,
    pub ticket_sequence: Option<u32>,
    pub txn_signature: Option<Cow<'a, str>>,

}

The base fields for all transaction models.

See Transaction Common Fields: <https://xrpl.org/transaction-common-fields.html>

Fields
account: Cow<'a, str>

The unique address of the account that initiated the transaction.

transaction_type: TransactionType

The type of transaction.

See Transaction Types: <https://xrpl.org/transaction-types.html>

account_txn_id: Option<Cow<'a, str>>

Hash value identifying another transaction. If provided, this transaction is only valid if the sending account’s previously-sent transaction matches the provided hash.

fee: Option<XRPAmount<'a>>

Integer amount of XRP, in drops, to be destroyed as a cost for distributing this transaction to the network. Some transaction types have different minimum requirements. See Transaction Cost for details.

flags: FlagCollection<F>

Set of bit-flags for this transaction.

last_ledger_sequence: Option<u32>

Highest ledger index this transaction can appear in. Specifying this field places a strict upper limit on how long the transaction can wait to be validated or rejected. See Reliable Transaction Submission for more details.

memos: Option<Vec<Memo>>

Additional arbitrary information used to identify this transaction.

network_id: Option<u32>

The network ID of the chain this transaction is intended for. MUST BE OMITTED for Mainnet and some test networks. REQUIRED on chains whose network ID is 1025 or higher.

sequence: Option<u32>

The sequence number of the account sending the transaction. A transaction is only valid if the Sequence number is exactly 1 greater than the previous transaction from the same account. The special case 0 means the transaction is using a Ticket instead.

signers: Option<Vec<Signer>>

Arbitrary integer used to identify the reason for this payment, or a sender on whose behalf this transaction is made. Conventionally, a refund should specify the initial payment’s SourceTag as the refund payment’s DestinationTag.

signing_pub_key: Option<Cow<'a, str>>

Hex representation of the public key that corresponds to the private key used to sign this transaction. If an empty string, indicates a multi-signature is present in the Signers field instead.

source_tag: Option<u32>

Arbitrary integer used to identify the reason for this payment, or a sender on whose behalf this transaction is made. Conventionally, a refund should specify the initial payment’s SourceTag as the refund payment’s DestinationTag.

ticket_sequence: Option<u32>

The sequence number of the ticket to use in place of a Sequence number. If this is provided, Sequence must be 0. Cannot be used with AccountTxnID.

txn_signature: Option<Cow<'a, str>>

The signature that verifies this transaction as originating from the account it says it is from.

Implementations

impl<'a, T> CommonFields<'a, T>
where T: IntoEnumIterator + Serialize + Debug,

pub fn new( account: Cow<'a, str>, transaction_type: TransactionType, account_txn_id: Option<Cow<'a, str>>, fee: Option<XRPAmount<'a>>, flags: Option<FlagCollection<T>>, last_ledger_sequence: Option<u32>, memos: Option<Vec<Memo>>, network_id: Option<u32>, sequence: Option<u32>, signers: Option<Vec<Signer>>, signing_pub_key: Option<Cow<'a, str>>, source_tag: Option<u32>, ticket_sequence: Option<u32>, txn_signature: Option<Cow<'a, str>>, ) -> Self

impl<T> CommonFields<'_, T>
where T: IntoEnumIterator + Serialize + Debug + PartialEq + Clone,

pub fn is_signed(&self) -> bool