Function xrpl::utils::parse_nftoken_id::unscramble_taxon

pub fn unscramble_taxon(taxon: u64, token_seq: u64) -> u64

Unscrambles or rescrambles a taxon in an NFTokenID.

An issuer may issue several NFTs with the same taxon; to ensure that NFTs are spread across multiple pages we lightly mix the taxon up by using the sequence (which is not under the issuer’s direct control) as the seed for a simple linear congruential generator.

From the Hull-Dobell theorem we know that f(x)=(m*x+c) mod n will yield a permutation of [0, n) when n is a power of 2 if m is congruent to 1 mod 4 and c is odd. By doing a bitwise XOR with this permutation we can scramble/unscramble the taxon.

The XLS-20d proposal fixes m = 384160001 and c = 2459. We then take the modulus of 2^32 which is 4294967296.